# docs/openapi-leanschool.yaml
openapi: 3.0.3
info:
  title: leanschool API
  description: CRUD data layer for school management (receipts, locations, persons,
    classes, etc.).
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: local
paths:
...
  /buildings:
    get:
      summary: List buildings
      operationId: listBuildings
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Building'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a building
      operationId: createBuilding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingRequest'
      responses:
        '201':
          description: Created Building
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /buildings/{id}:
    get:
      summary: Get a Building by ID
      operationId: getBuilding
      parameters:
      - name: id
        in: path
        required: true
        description: Building ID
        schema:
          type: string
      responses:
        '200':
          description: Building
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '404':
          description: Not found
    put:
      summary: Update a Building
      operationId: updateBuilding
      parameters:
      - name: id
        in: path
        required: true
        description: Building ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingRequest'
      responses:
        '200':
          description: Updated Building
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a Building
      operationId: deleteBuilding
      parameters:
      - name: id
        in: path
        required: true
        description: Building ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /cities:
    get:
      summary: List cities
      operationId: listCitys
      responses:
        '200':
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a citie
      operationId: createCity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityRequest'
      responses:
        '201':
          description: Created City
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /cities/{id}:
    get:
      summary: Get a City by ID
      operationId: getCity
      parameters:
      - name: id
        in: path
        required: true
        description: City ID
        schema:
          type: string
      responses:
        '200':
          description: City
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '404':
          description: Not found
    put:
      summary: Update a City
      operationId: updateCity
      parameters:
      - name: id
        in: path
        required: true
        description: City ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityRequest'
      responses:
        '200':
          description: Updated City
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a City
      operationId: deleteCity
      parameters:
      - name: id
        in: path
        required: true
        description: City ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /exams:
    get:
      summary: List exams
      operationId: listExams
      responses:
        '200':
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exam'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a exam
      operationId: createExam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamRequest'
      responses:
        '201':
          description: Created Exam
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /exams/{id}:
    get:
      summary: Get a Exam by ID
      operationId: getExam
      parameters:
      - name: id
        in: path
        required: true
        description: Exam ID
        schema:
          type: string
      responses:
        '200':
          description: Exam
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '404':
          description: Not found
    put:
      summary: Update a Exam
      operationId: updateExam
      parameters:
      - name: id
        in: path
        required: true
        description: Exam ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamRequest'
      responses:
        '200':
          description: Updated Exam
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a Exam
      operationId: deleteExam
      parameters:
      - name: id
        in: path
        required: true
        description: Exam ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  schemas:
    Address:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: string
          example: 1a
        postalCode:
          $ref: '#/components/schemas/PostalCode'
        street:
          type: string
          example: Bahnhofstrasse
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    Building:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        location:
          $ref: '#/components/schemas/Location'
        name:
          type: string
          example: Main Building
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - version
    BuildingRequest:
      type: object
      properties:
        location:
          $ref: '#/components/schemas/EntityRef'
        name:
          type: string
          example: Main Building
        version:
          type: integer
          example: 0
      required:
      - name
      - version
    City:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Zurich
        postalCodes:
          type: array
          items:
            $ref: '#/components/schemas/PostalCode'
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - postalCodes
      - version
    CityRequest:
      type: object
      properties:
        name:
          type: string
          example: Zurich
        postalCodes:
          type: array
          items:
            $ref: '#/components/schemas/PostalCodeRef'
        version:
          type: integer
          example: 0
      required:
      - name
      - postalCodes
      - version
...
    EntityRef:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      required:
      - id
    Exam:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        schoolClass:
          $ref: '#/components/schemas/SchoolClass'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    ExamRequest:
      type: object
      properties:
        schoolClass:
          $ref: '#/components/schemas/EntityRef'
        version:
          type: integer
          example: 0
      required:
      - version
    Grade:
      type: object
      properties:
        exam:
          $ref: '#/components/schemas/Exam'
        grade:
          type: number
          format: double
          example: 5.5
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        student:
          $ref: '#/components/schemas/Student'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    Guardian:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - version
...
    Location:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        lat:
          type: number
          format: double
          example: 47.3769
        lon:
          type: number
          format: double
          example: 8.5417
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    PostalCode:
      type: object
      properties:
        city:
          type: string
          example: Zurich
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: integer
          example: 8001
        version:
          type: integer
          example: 0
      required:
      - id
      - city
      - version
      - number
    PostalCodeRef:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      required:
      - id
...
    Room:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/Building'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Room 101
        version:
          type: integer
          example: 0
        roomType:
          $ref: '#/components/schemas/RoomType'
      required:
      - id
      - name
      - version
...
    SchoolClass:
      type: object
      properties:
        classroom:
          $ref: '#/components/schemas/Room'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Class 5A
        schoolYear:
          $ref: '#/components/schemas/SchoolYear'
        shortcut:
          type: string
          example: 5A
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/Teacher'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
      - name
      - shortcut
      - teachers
      - students
...
    SchoolYear:
      type: object
      properties:
        from:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        to:
          type: string
          format: date-time
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    Student:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        currentSchoolClass:
          $ref: '#/components/schemas/SchoolClass'
        dateOfBirth:
          type: string
          format: date-time
        grades:
          type: array
          items:
            $ref: '#/components/schemas/Grade'
        guardians:
          type: array
          items:
            $ref: '#/components/schemas/Guardian'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Doe
        pastSchoolClasses:
          type: array
          items:
            $ref: '#/components/schemas/SchoolClass'
        prename:
          type: string
          example: Jane
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - guardians
      - version
      - grades
      - pastSchoolClasses
...
    Teacher:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        atSchoolSince:
          type: string
          format: date-time
        atSchoolUntil:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        iban:
          type: string
          example: CH93-0076-2011-6238-5295-7
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - atSchoolSince
      - version
...
    RoomType:
      type: string
      enum:
      - ClassRoom
      - Kindergarden
      - Music
      - Video
      - Library
      - TeacherRoom
      - ComputerRoom
      - FacilityManagementRoom
      - Theater
      - MeetingRoom
      - ScienceRoom
      - SchoolManagementRoom
      - SchoolSecretariatRoom
...
  securitySchemes:
    keycloak:
      type: openIdConnect
      description: Keycloak OpenID Connect — JWT bearer token with realm_access roles
        (leanschool_read, leanschool_write)
      openIdConnectUrl: http://localhost:8180/realms/leanschool/.well-known/openid-configuration
security:
- keycloak: []